// src/enum.ts
export enum ERROR_CODES {
  HTTP_BAD_REQUEST = 400,
  HTTP_NOT_AUTHORIZED = 403,
  HTTP_NOT_FOUND = 404,
  HTTP_SERVER_ERROR = 500,
}

export enum ERROR_MESSAGES {
  HTTP_BAD_REQUEST = "Bad request",
  HTTP_NOT_AUTHORIZED = "User not authorized",
  HTTP_NOT_FOUND = "Resource not found",
  HTTP_SERVER_ERROR = "Internal server error",
}